refactor(frontend): consolidate wire-contract payload types under types/payloads
Move ActivityPayload, CommentPayload, SuggestionPayload, OwnershipPayload,
AgentPresencePayload, and LinkAccess out of page/component/editor modules
into one types/payloads.ts that mirrors the Rails as_props serializers.
Type the author/actor kind and status fields as unions instead of string,
and fix the components -> pages import inversion in activity_panel.
