+++ app/frontend/components/margin_suggestions.tsx
@@ -5,7 +5,8 @@ import { TextSelection } from '@milkdown/kit/prose/state'
 import type { EditorView } from '@milkdown/kit/prose/view'
 import type { EditorHandle } from '../editor/milkdown_editor'
 import type { ProvenanceSpan } from '../editor/provenance'
-import { findSuggestionTarget, type SuggestionPayload } from '../editor/suggestions'
+import { findSuggestionTarget } from '../editor/suggestions'
+import type { SuggestionPayload } from '../types/payloads'
 import { truncate } from '../lib/truncate'
 import { clearHighlight, domRange, setHighlight, supportsHighlights } from '../lib/highlights'
 import { sourceParser, type DocumentFormat } from '../editor/document_format'
